+++ Editor/OptimizerPlugin.cs
@@ -75,6 +75,8 @@ namespace Anatawa12.AvatarOptimizer.ndmf
                         ;
                 });
 
+#if UNITY_2021_3_OR_NEWER
+            // animator optimizer is written in newer C# so requires 2021.3 or newer 
             mainSequence.Run(Processors.AnimatorOptimizer.InitializeAnimatorOptimizer.Instance)
 #if AAO_VRCSDK3_AVATARS
                 // EntryExit to BlendTree optimization heavily depends on VRChat's behavior
@@ -83,6 +85,7 @@ namespace Anatawa12.AvatarOptimizer.ndmf
                 .Then.Run(Processors.AnimatorOptimizer.MergeDirectBlendTree.Instance)
                 .Then.Run(Processors.AnimatorOptimizer.RemoveMeaninglessLayer.Instance)
                 ;
+#endif
         }
 
         protected override void OnUnhandledException(Exception e)