+++ Internal/Localization/Editor/BuildLog.cs
@@ -1,0 +1,16 @@
+using nadena.dev.ndmf;
+
+namespace Anatawa12.AvatarOptimizer
+{
+    public static class BuildLog
+    {
+        public static void LogInfo(string code, params object[] args) =>
+            ErrorReport.ReportError(AAOL10N.Localizer, ErrorSeverity.Information, code, args);
+
+        public static void LogWarning(string code, params object[] args) =>
+            ErrorReport.ReportError(AAOL10N.Localizer, ErrorSeverity.NonFatal, code, args);
+
+        public static void LogError(string code, params object[] args) =>
+            ErrorReport.ReportError(AAOL10N.Localizer, ErrorSeverity.Error, code, args);
+    }
+}