+++ Test~/Basic/BuildAssetBundle.cs
@@ -1,0 +1,29 @@
+using System.IO;
+using NUnit.Framework;
+using UnityEditor;
+
+namespace Anatawa12.AvatarOptimizer.Test
+{
+    /// <summary>
+    /// This tests compilation error in runtime build with building asset bundle.
+    /// </summary>
+    public class BuildAssetBundle
+    {
+        [Test]
+        public void Build()
+        {
+            BuildPipeline.BuildAssetBundles("Assets/",
+                new[]
+                {
+                    new AssetBundleBuild
+                    {
+                        assetNames = new[] { TestUtils.GetAssetPath("Basic/Empty.prefab") },
+                        assetBundleName = "asset.unity3d"
+                    }
+                },
+                BuildAssetBundleOptions.None,
+                EditorUserBuildSettings.activeBuildTarget);
+        }
+    }
+}
+